Use correct namespace, ensure spec helpers have all needed methods
Legacy AR gem depends now on current AR gem.
