+++ rails_event_store_active_record-legacy/lib/rails_event_store_active_record/legacy/generators/templates/v1_v2_migration_template.rb
