+++ rails_event_store_active_record-legacy/rails_event_store_active_record-legacy.gemspec
@@ -35,10 +35,12 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'sqlite3', '1.3.13'
   spec.add_development_dependency 'pg', '0.21'
   spec.add_development_dependency 'mysql2', '0.4.10'
+  spec.add_development_dependency 'fakefs', '~> 0.11.2'
   spec.add_development_dependency 'childprocess'
   spec.add_development_dependency 'mutant-rspec', '~> 0.8.15'
 
   spec.add_dependency 'ruby_event_store', '= 0.31.1'
+  spec.add_dependency 'rails_event_store_active_record', '= 0.31.1'
   spec.add_dependency 'activesupport', '>= 3.0'
   spec.add_dependency 'activemodel', '>= 3.0'
 end
