+++ rails_event_store_active_record-legacy/spec/v1_v2_migration_generator_spec.rb
@@ -3,37 +3,39 @@ require 'pp'
 require 'fakefs/safe'
 
 module RailsEventStoreActiveRecord
-  RSpec.describe V1V2MigrationGenerator do
-    around(:each) do |example|
-      current_stdout = $stdout
-      $stdout = StringIO.new
-      example.call
-      $stdout = current_stdout
-    end
-
-    specify do
-      FakeFS.with_fresh do
-        FakeFS::FileSystem.clone(File.expand_path('../../', __FILE__))
-        stub_const("Rails::VERSION::STRING", "4.2.8")
-
-        generator = V1V2MigrationGenerator.new
-        allow(Time).to receive(:now).and_return(Time.new(2016,8,9,22,22,22))
-        generator.create_migration
-
-        expect(File.read("db/migrate/20160809222222_migrate_res_schema_v1_to_v2.rb")).to match(/ActiveRecord::Migration$/)
+  module Legacy
+    RSpec.describe V1V2MigrationGenerator do
+      around(:each) do |example|
+        current_stdout = $stdout
+        $stdout = StringIO.new
+        example.call
+        $stdout = current_stdout
       end
-    end
 
-    specify do
-      FakeFS.with_fresh do
-        FakeFS::FileSystem.clone(File.expand_path('../../', __FILE__))
-        stub_const("Rails::VERSION::STRING", "5.0.0")
+      specify do
+        FakeFS.with_fresh do
+          FakeFS::FileSystem.clone(File.expand_path('../../', __FILE__))
+          stub_const("Rails::VERSION::STRING", "4.2.8")
 
-        generator = V1V2MigrationGenerator.new
-        allow(Time).to receive(:now).and_return(Time.new(2016,8,9,22,22,22))
-        generator.create_migration
+          generator = V1V2MigrationGenerator.new
+          allow(Time).to receive(:now).and_return(Time.new(2016,8,9,22,22,22))
+          generator.create_migration
 
-        expect(File.read("db/migrate/20160809222222_migrate_res_schema_v1_to_v2.rb")).to match(/ActiveRecord::Migration\[4\.2\]$/)
+          expect(File.read("db/migrate/20160809222222_migrate_res_schema_v1_to_v2.rb")).to match(/ActiveRecord::Migration$/)
+        end
+      end
+
+      specify do
+        FakeFS.with_fresh do
+          FakeFS::FileSystem.clone(File.expand_path('../../', __FILE__))
+          stub_const("Rails::VERSION::STRING", "5.0.0")
+
+          generator = V1V2MigrationGenerator.new
+          allow(Time).to receive(:now).and_return(Time.new(2016,8,9,22,22,22))
+          generator.create_migration
+
+          expect(File.read("db/migrate/20160809222222_migrate_res_schema_v1_to_v2.rb")).to match(/ActiveRecord::Migration\[4\.2\]$/)
+        end
       end
     end
   end
