+++ rails_event_store_active_record-legacy/spec/v1_v2_schema_migration_spec.rb
@@ -4,6 +4,7 @@ require 'childprocess'
 require 'active_record'
 require 'logger'
 require 'ruby_event_store'
+require 'rails_event_store_active_record'
 require 'ruby_event_store/spec/event_repository_lint'
 
 class EventA2 < RubyEventStore::Event
@@ -12,7 +13,7 @@ end
 RSpec.describe "v1_v2_migration" do
   include SchemaHelper
 
-  MigrationRubyCode = File.read(File.expand_path('../../lib/rails_event_store_active_record/generators/templates/v1_v2_migration_template.rb', __FILE__) )
+  MigrationRubyCode = File.read(File.expand_path('../../lib/rails_event_store_active_record/legacy/generators/templates/v1_v2_migration_template.rb', __FILE__) )
   migration_version = Gem::Version.new(ActiveRecord::VERSION::STRING) < Gem::Version.new("5.0.0") ? "" : "[4.2]"
   MigrationRubyCode.gsub!("<%= migration_version %>", migration_version)
