+++ rails_event_store_active_record/lib/rails_event_store_active_record.rb
@@ -1,5 +1,4 @@
 require 'rails_event_store_active_record/generators/migration_generator'
-require 'rails_event_store_active_record/generators/v1_v2_migration_generator'
 require 'rails_event_store_active_record/event'
 require 'rails_event_store_active_record/event_repository'
 require 'rails_event_store_active_record/event_repository_reader'
